Time: 0 ms (100.00%), Space: 40.6 MB (84.40%) - LeetHub

+++ 0050-powx-n/0050-powx-n.java
@@ -1,6 +1,6 @@
 class Solution {
     public double myPow(double x, int n) {
         
-        return Math.pow(x,n);
+        return (double)Math.pow(x,n);
     }
 }
